refactor(knowledge): organize retrieval backends
